requirements.txt
<<<<<<< HEAD
streamlit
openai
langchain
langchain-community
faiss-cpu
easyocr
openai-whisper
sympy
python-dotenv
numpy
=======
streamlit
openai
langchain
langchain-community
faiss-cpu
easyocr
openai-whisper
sympy
python-dotenv
numpy
>>>>>>> f67757f31ab44a651e38f5373c7b21bc1ba10434
pandas